backend/requirements.txt
fastapi
fastapi-login 
oauthlib 
requests-oauthlib
uvicorn[standard]
alembic
sqlmodel
sqlalchemy
pymysql
cryptography
argon2-cffi                                                                                                                                                                                                          
fastapi-mail
pyotp
python-multipart
redis
python-jose[cryptography]
passlib
faker
